Phase 1: Move OutboxRelayLoader + UsingRelayUnwrapper to commons
Move these two topNavFeeds classes to commons module, replacing
LocalCache parameter with ICacheProvider. Leave typealiases at
original locations for backward compatibility.
