+++ amethyst/src/main/java/com/vitorpamplona/amethyst/model/topNavFeeds/UsingRelayUnwrapper.kt
@@ -20,96 +20,4 @@
  */
 package com.vitorpamplona.amethyst.model.topNavFeeds
 
-import com.vitorpamplona.amethyst.model.AddressableNote
-import com.vitorpamplona.amethyst.model.LocalCache
-import com.vitorpamplona.amethyst.model.NoteState
-import com.vitorpamplona.quartz.nip01Core.core.HexKey
-import com.vitorpamplona.quartz.nip01Core.relay.normalizer.NormalizedRelayUrl
-import com.vitorpamplona.quartz.nip17Dm.settings.ChatMessageRelayListEvent
-import com.vitorpamplona.quartz.nip65RelayList.AdvertisedRelayListEvent
-import com.vitorpamplona.quartz.utils.mapOfSet
-import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.combine
-
-class UsingRelayUnwrapper {
-    fun usersPerRelay(relayListNotes: Array<NoteState>): Map<NormalizedRelayUrl, Set<HexKey>> =
-        mapOfSet {
-            relayListNotes.forEach { outboxNote ->
-                val note = outboxNote.note
-
-                val authorHex =
-                    if (note is AddressableNote) {
-                        note.address.pubKeyHex
-                    } else {
-                        note.author?.pubkeyHex
-                    }
-
-                if (authorHex != null) {
-                    val relays =
-                        when (val noteEvent = outboxNote.note.event) {
-                            is AdvertisedRelayListEvent -> noteEvent.relaysNorm()
-                            is ChatMessageRelayListEvent -> noteEvent.relays()
-                            else -> emptySet()
-                        }
-
-                    relays.forEach {
-                        add(it, authorHex)
-                    }
-                    relays.forEach {
-                        add(it, authorHex)
-                    }
-                }
-            }
-        }
-
-    fun <T> usersPerRelaySnapshot(
-        users: Set<HexKey>,
-        cache: LocalCache,
-        transformation: (Map<NormalizedRelayUrl, Set<HexKey>>) -> T,
-    ): T {
-        val nip65RelayNotes =
-            users
-                .map { pubkeyHex ->
-                    cache
-                        .getOrCreateAddressableNote(AdvertisedRelayListEvent.createAddress(pubkeyHex))
-                        .flow()
-                        .metadata.stateFlow.value
-                }.toTypedArray()
-
-        val nip17RelayNotes =
-            users
-                .map { pubkeyHex ->
-                    cache
-                        .getOrCreateAddressableNote(ChatMessageRelayListEvent.createAddress(pubkeyHex))
-                        .flow()
-                        .metadata.stateFlow.value
-                }.toTypedArray()
-
-        return transformation(usersPerRelay(nip65RelayNotes + nip17RelayNotes))
-    }
-
-    fun <T> toUsersPerRelayFlow(
-        users: Set<HexKey>,
-        cache: LocalCache,
-        transformation: (Map<NormalizedRelayUrl, Set<HexKey>>) -> T,
-    ): Flow<T> {
-        val relayNoteFlows =
-            users.map { pubkeyHex ->
-                val note = cache.getOrCreateAddressableNote(AdvertisedRelayListEvent.createAddress(pubkeyHex))
-                note.flow().metadata.stateFlow
-            } +
-                users.map { pubkeyHex ->
-                    val note = cache.getOrCreateAddressableNote(ChatMessageRelayListEvent.createAddress(pubkeyHex))
-                    note.flow().metadata.stateFlow
-                }
-
-        return if (relayNoteFlows.isEmpty()) {
-            MutableStateFlow(transformation(emptyMap()))
-        } else {
-            combine(relayNoteFlows) { relayListNotes ->
-                transformation(usersPerRelay(relayListNotes))
-            }
-        }
-    }
-}
+typealias UsingRelayUnwrapper = com.vitorpamplona.amethyst.commons.model.topNavFeeds.UsingRelayUnwrapper
